--- functions/config/db.js
@@ -1,8 +1,0 @@
-const admin = require('firebase-admin');
-require('firebase/firestore');
-
-admin.initializeApp();
-
-const db = admin.firestore();
-
-module.exports = db;
