--- functions/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "functions",
-  "description": "Cloud Functions for Firebase",
-  "engines": {
-    "node": "10"
-  },
-  "scripts": {
-    "lint": "eslint .",
-    "serve": "firebase serve --only functions",
-    "shell": "firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "dependencies": {
-    "firebase": "^7.3.0",
-    "firebase-admin": "^8.7.0",
-    "firebase-functions": "^3.3.0",
-    "moment": "^2.24.0",
-    "request": "^2.88.0",
-    "request-promise-native": "^1.0.8"
-  },
-  "devDependencies": {
-    "eslint": "^5.12.0",
-    "eslint-plugin-promise": "^4.0.1"
-  }
-}
