--- legacy/config/constants.js
@@ -1,17 +1,0 @@
-module.exports = { //Constants used for logging & responsmessages
-  DATABASE_ERROR: 'DATABASE_ERROR',
-  GENERAL_ERROR: 'GENERAL_ERROR',
-  INVALID_REQUEST_ERROR: 'INVALID_REQUEST_ERROR',
-  MISSING_REQUEST_ERROR: 'MISSING_REQUEST_ERROR',
-  NETWORK_CALL_ERROR: 'NETWORK_CALL_ERROR',
-  NO_MATCH: 'NO_MATCH',
-  NO_RESULTS: 'NO_RESULTS',
-  NOT_ACCEPTING: 'NOT_ACCEPTING',
-  SUCCESS: 'SUCCESS',
-  TOO_MANY_GIFTS: 'TOO_MANY_GIFTS',
-  USER_DOES_NOT_EXIST: 'USER_DOES_NOT_EXIST',
-  USER_EXISTS: 'USER_EXISTS',
-  WRONG_API_TOKEN: 'WRONG_API_TOKEN',
-  WRONG_API_TOKEN_NAME: 'WRONG_API_TOKEN_NAME',
-  UPDATE_SUCCESS: 'UPDATE_SUCCESS'
-};
