--- legacy/routes/appRoutes.js
@@ -1,6 +1,0 @@
-const express = require('express');
-let router = express.Router();
-
-router.route('/');
-
-module.exports = router;
